refactor: create src folder for electoral college
