+++ electoral_college/enums.py
