+++ electoral_college/exceptions.py
