+++ electoral_college/vote_calculator.py
